Restructures DashboardAdmin folder
